+++ werobot/config.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+
+class Config(dict):
+    pass